Feature: Added product page for specific Category
It Includes feature to view product card in card view or in list view.
It also includes the filter component that manages the specific category and in
the right we can switch to different category by the help of right hand
side panel.
I have created the loading icon which appear when we are fetching the
data from the api and setting them into the page.

I have created the card components which takes the products data as the
prop and show you in the card or list view.

In the utils folder,
I have created models file which include two functions
one for connecting to the database with the help of environment variables
and second is for creating model and model is already created then it
will return the already created model.
I have also created the schemas file which include the schemas for the
database collections and it is also the dependency of the models file.

I have created the environment variables file .env.development.

I have created the api for connecting to database and fetching data for
specific products but there is a problem while fetching the data.